Build out skeleton rendering users, search bar
To stop the infinite API calls, I've set up another useState() hook which is called right after setData() that sets the first user's name to its value. This is a bit of a hacky way to stop the API calls after the first re-render (two API calls are made in total). Will come back to this issue if need be.