Add album name with each Photo entry
Also added a janky "pagination" that involves showing more photos when a "Show More" button is clicked. This should be changed.

Still concerned with the time it takes to make each API call; going to look into cleaning up the "PhotoList" component to reduce the work done to render the list of photos for each user.