Add search functionality and delete Search component
Initially wanted a separate Search component, but realized that since the searched users are being rendered through UserList, it might be betterto build the search functionality into App and pass the changed users directly into the UserList component.

Search functionality is handled by handleSearch(), which takes in a string term and compares it with the existing users array. This implementation creates duplicate state (renderedUsers and users), for which I will take into account as a road block.