Add Search component and patch type error in Paginate component
Search component now holds all search logic that was previously held in App.

Paginate component ran into issue that when selecting an option in the page dropdown, the value's type would automatically convert to string. As such, this value is converted into a number whenever it is passed into function setCurrentPage(), which changes the currentPage state in PhotoList.